Moved pcap file manipulation functions to appropriate source code
